requirements.txt
aiogram==3.24.0
aiohttp==3.13.3
python-dotenv==1.2.1
aiosqlite==0.22.1
prometheus-client>=0.20.0
fastapi>=0.110.0
uvicorn>=0.27.0
